+++ packages/admin-rte/src/core/Controls/useBlockTypes.tsx
@@ -39,7 +39,7 @@ const createFeaturesFromBlocktypeMap = (group: "dropdown" | "button") => ({
             label: config.label ?? blocktype,
             selected: blockTypeActive(blocktype),
             onButtonClick: handleBlockTypeButtonClick.bind(null, blocktype),
-            Icon: config.icon,
+            icon: config.icon,
         })),
 ];