+++ packages/admin-rte/CHANGELOG.md
@@ -13,10 +13,11 @@ All notable changes to this project will be documented in this file. This projec
 ### Internal Changes
 
 -   Rename prop-name "customBlockMap" to "blocktypeMap", deprecate prop-name "customBlockMap"
+-   Rename prop-name "Icon" to "icon", deprecate prop-name "Icon"
+
 ## [1.1.1] - 12. Jan 2021 - re-release under new name
 
 This package has been renamed to @comet/admin-rte
-
 ## [1.1.1] - 11. Jan 2021
 
 This is a bugfix/maintenance release